+++ core/commons/object_labels_providers/enclave_object_labels_provider.go
@@ -11,6 +11,8 @@ import (
 	"github.com/kurtosis-tech/kurtosis-core/api_container/server/service_network/service_network_types"
 	"github.com/kurtosis-tech/kurtosis-core/commons/enclave_object_labels"
 	"net"
+	"strconv"
+	"time"
 )
 
 type EnclaveObjectLabelsProvider struct {
@@ -22,7 +24,7 @@ func NewEnclaveObjectLabelsProvider(enclaveId string) *EnclaveObjectLabelsProvid
 }
 
 func (labelsProvider *EnclaveObjectLabelsProvider) ForApiContainer(apiContainerIPAddress net.IP, apiContainerListenPort uint16) map[string]string {
-	labels := labelsProvider.getLabelsMapWithCommonsLabels()
+	labels := labelsProvider.getLabelsMapForEnclaveObject()
 	labels[enclave_object_labels.ContainerTypeLabel] = enclave_object_labels.ContainerTypeAPIContainer
 	labels[enclave_object_labels.APIContainerURLLabel] = fmt.Sprintf("%v:%v", apiContainerIPAddress.String(), apiContainerListenPort)
 	return labels
@@ -31,37 +33,46 @@ func (labelsProvider *EnclaveObjectLabelsProvider) ForApiContainer(apiContainerI
 // TODO We don't want testsuites to be special - they should be Just Another Kurtosis Module - but we can't make them
 //  unspecial (and thus delete this method) until the API container supports a container log-streaming endpoint
 func (labelsProvider *EnclaveObjectLabelsProvider) ForTestRunningTestsuiteContainer() map[string]string {
-	labels := labelsProvider.getLabelsMapWithCommonsLabels()
+	labels := labelsProvider.getLabelsMapForEnclaveObject()
 	labels[enclave_object_labels.ContainerTypeLabel] = enclave_object_labels.ContainerTypeTestsuiteContainer
 	return labels
 }
 
 func (labelsProvider *EnclaveObjectLabelsProvider) ForUserServiceContainer(serviceGUID service_network_types.ServiceGUID) map[string]string {
-	labels := labelsProvider.getLabelsMapWithCommonsLabelsAndGUIDLabel(string(serviceGUID))
+	labels := labelsProvider.getLabelsMapForEnclaveObjectWithGUID(string(serviceGUID))
 	labels[enclave_object_labels.ContainerTypeLabel] = enclave_object_labels.ContainerTypeUserServiceContainer
 	return labels
 }
 
 func (labelsProvider *EnclaveObjectLabelsProvider) ForNetworkingSidecarContainer(serviceGUID service_network_types.ServiceGUID) map[string]string {
-	labels := labelsProvider.getLabelsMapWithCommonsLabelsAndGUIDLabel(string(serviceGUID))
+	labels := labelsProvider.getLabelsMapForEnclaveObjectWithGUID(string(serviceGUID))
 	labels[enclave_object_labels.ContainerTypeLabel] = enclave_object_labels.ContainerTypeNetworkingSidecarContainer
 	return labels
 }
 
 func (labelsProvider *EnclaveObjectLabelsProvider) ForLambdaContainer(lambdaGUID lambda_store_types.LambdaGUID) map[string]string {
-	labels := labelsProvider.getLabelsMapWithCommonsLabelsAndGUIDLabel(string(lambdaGUID))
+	labels := labelsProvider.getLabelsMapForEnclaveObjectWithGUID(string(lambdaGUID))
 	labels[enclave_object_labels.ContainerTypeLabel] = enclave_object_labels.ContainerTypeLambdaContainer
 	return labels
 }
 
-func (labelsProvider *EnclaveObjectLabelsProvider) getLabelsMapWithCommonsLabels() map[string]string {
+// This is a liiiiittle strange to have here, since this is only used by the CLI (not anything inside this repo)
+func (labelsProvider *EnclaveObjectLabelsProvider) ForInteractiveREPLContainer(interactiveReplTimestampGuid time.Time) map[string]string {
+	// TODO Format this as base-16 (hex) to make this more concise??
+	interactiveReplTimestampUnixSecsStr := strconv.FormatInt(interactiveReplTimestampGuid.Unix(), 10)
+	labels := labelsProvider.getLabelsMapForEnclaveObjectWithGUID(interactiveReplTimestampUnixSecsStr)
+	labels[enclave_object_labels.ContainerTypeLabel] = enclave_object_labels.ContainerTypeInteractiveREPL
+	return labels
+}
+
+func (labelsProvider *EnclaveObjectLabelsProvider) getLabelsMapForEnclaveObject() map[string]string {
 	labels := map[string]string{}
 	labels[enclave_object_labels.EnclaveIDContainerLabel] = labelsProvider.enclaveId
 	return labels
 }
 
-func (labelsProvider *EnclaveObjectLabelsProvider) getLabelsMapWithCommonsLabelsAndGUIDLabel(guid string) map[string]string {
-	labels := labelsProvider.getLabelsMapWithCommonsLabels()
+func (labelsProvider *EnclaveObjectLabelsProvider) getLabelsMapForEnclaveObjectWithGUID(guid string) map[string]string {
+	labels := labelsProvider.getLabelsMapForEnclaveObject()
 	labels[enclave_object_labels.GUIDLabel] = guid
 	return labels
 }