+++ core/docs/changelog.md
@@ -1,4 +1,8 @@
 # TBD
+### Features
+* Added a new `interactive-repl` value to the `container-type` label
+* Added `ForInteractiveREPLContainer` functions to the `EnclaveObjectLabelsProvider` and `EnclaveObjectNameProvider`
+* Added a `GetContainerGUIDSuffix` function to centralize generation of container GUID suffixes
 
 # 1.23.3
 ### Features